+++ package.json
@@ -11,16 +11,9 @@
     "cypress:open": "TZ=America/Los_Angeles cypress open",
     "plugin_helpers": "node ../../scripts/plugin_helpers"
   },
-  "dependencies": {
-    "react-grid-layout": "^0.16.2"
-  },
+  "dependencies": {},
   "devDependencies": {
-    "@cypress/skip-test": "^2.6.1",
-    "@types/enzyme-adapter-react-16": "^1.0.6",
-    "@types/react-test-renderer": "^16.9.1",
-    "@types/react-grid-layout": "^0.16.7",
-    "cypress": "^5.0.0",
-    "eslint": "^6.8.0",
-    "jest-dom": "^4.0.0"
+    "cypress": "^10.11.0",
+    "eslint": "^6.8.0"
   }
 }
