+++ public/constants/index.ts
@@ -2,8 +2,3 @@
  * Copyright OpenSearch Contributors
  * SPDX-License-Identifier: Apache-2.0
  */
-
-export enum ServiceEndpoints {
-  GetIndexes = '/api/relevancy/search/indexes',
-  GetSearchResults = '/api/relevancy/search',
-}
